Added multiple image galleries, added GIF support to gallery thumbnailer, fixed renaming bug in massren.sh, misc. fixes in other scripts.

+++ tools/gallerygen/directory.sh
@@ -5,7 +5,19 @@ orange=`tput setaf 3`
 cyan=`tput setaf 6`
 reset=`tput sgr0`
 
+function get_image_info {
+	NAME="${f%%.*}"
+	EXTENSION="${f##*.}"
+	OUTEXT="jpg"
+	THUMBOUTPATH="../../out/$GALLERY/thumbnail_$NAME.$OUTEXT"
+	
+	if [[ "$EXTENSION" == "gif" ]]; then
+		THUMBOUTPATH="../../out/$GALLERY/thumbnail_$NAME.gif"
+		OUTEXT="gif"
+	fi
 
+	OUTPATH="../../out/$GALLERY/$f"
+} 
 
 if [[ "$#" < 1 ]]; then
 	echo "not enough args"
@@ -20,27 +32,47 @@ else
 	mkdir -p out/"$GALLERY"
 
 	cd $1
+	FILES=$(ls *.{gif,png} | sort | tr '\n' ' ')	
+	
+	# Avoid needless processing
+
+	SKIPDIRECTORY=true
+
+	for f in $FILES; do
+		get_image_info
+		if [[ $f -nt $THUMBOUTPATH || ! -f $THUMBOUTPATH ]]; then
+			SKIPDIRECTORY=false
+			break;
+		fi
+	done
+		
+	if [[ "$SKIPDIRECTORY" = true ]]; then
+		echo Skipping.
+		exit 0
+	fi
+
+	##
 
 	rm "$METADATA" 2> /dev/null
 
-	for f in *.png; do
-		NAME=$(echo "$f" | sed 's/\./ /g' | cut -d' ' -f1);
-		EXTENSION=$(echo "$f" | sed 's/\./ /g' | cut -d' ' -f2);
-		THUMBOUTPATH="../../out/$GALLERY/thumbnail_$NAME.jpg"
-		OUTPATH="../../out/$GALLERY/$f"
-		
+	for f in $FILES; do
+		get_image_info		
 		DESCRIPTION="$(magick identify -format "%[Description]" $f 2>/dev/null)"
 		
-
 		if [[ $f -nt $THUMBOUTPATH || ! -f $THUMBOUTPATH ]]; then
 			echo "$green""Converting $f -> $THUMBOUTPATH$reset"
-			magick convert -strip -thumbnail '240x140^' "$f" -quality 100 "$THUMBOUTPATH"
+			if [[ "$EXTENSION" == "gif" ]]; then
+				magick convert -strip -thumbnail '240x140^' "$f" -layers Optimize +map "$THUMBOUTPATH"
+				gifsicle -U -O3 --colors=32 --dither=ordered "$THUMBOUTPATH" -o "$THUMBOUTPATH"
+			else
+				magick convert -strip -thumbnail '240x140^' "$f[0]" -quality 100 "$THUMBOUTPATH"
+			fi
 		fi
 
 		cp -R -u "$f" "$OUTPATH"
 
 		echo "- image: /assets/galleries/$GALLERY/$f" >> $METADATA
-		echo "  thumbnail: /assets/galleries/$GALLERY/thumbnail_$NAME.jpg" >> $METADATA
+		echo "  thumbnail: /assets/galleries/$GALLERY/thumbnail_$NAME.$OUTEXT" >> $METADATA
 		if [ ! -z "$DESCRIPTION" ]; then
 			echo "  description: \"$DESCRIPTION\"" >> $METADATA
 		fi
